(Issue #3) : Create DelegateCommandBase class. DelegateCommand class's some method move to DelegateCommandBase class and inherit DelegateCommandBase .
